refactor: Redis 관련 설정 파일들을 infra 계층으로 이동 - domain 계층에 위치해 있던 RedissonConfig, RedisProperties 등을 infra.redis 또는 infra.lock 등으로 이동 - 관심사 분리를 위해 아키텍처 구조 정리
