+++ src/main/java/io/hhplus/concert/infrastructure/config/RedisConfig.java
@@ -1,6 +1,6 @@
-package io.hhplus.concert.domain.user;
+package io.hhplus.concert.infrastructure.config;
 
-import io.hhplus.concert.domain.reservation.RedisProperties;
+import io.hhplus.concert.infrastructure.config.RedisProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
